--- src/Context/useTask.tsx
@@ -1,10 +1,0 @@
-import { useContext } from "react";
-import { TaskContext } from "../Context/TaskContext";
-
-export function useTasks() {
-  const context = useContext(TaskContext);
-  if (!context) {
-    throw new Error("useTasks must be used within a TaskProvider");
-  }
-  return context;
-}
